Patch prune_binaries.py to not exit if targets for pruning are missing If setup has been run on a prior occasion, then prune_binaries.py exits because if it cannot find the already-pruned files.
This seems reasonable when testing pruning.list for files which no longer
exist in the chromium tarball, but in our case prevents setup being re-run
(at least in the case that setup has advanced to the pruning step and
resulted in file deletion).

This is probably not the right way to fix this, as the change might wrongly
allow things to continue in the event of a permission error. That said, I
cannot ever recall getting a permission error.