Fix build not being optimised and rely on upstream optimisation Confirmed by adding -v to the ninja command in rules, -O2 wasn't getting added to the clang command line, probably because we're not using the unbundled toolchain.
Remove use-package-optimizations.patch and rely on the upstream
optimisation decisions, which are probably a bit smarter in that they
pick and choose optimisation levels depending on the target (presumably to
get a good trade-off between size and performance).

With pgo and sample_profile_is_accurate=true, a fair bit of code will
get optimised for size (ie -Oz).

In addition, remove the setting of CXXFLAGS and CFLAGS in rules, as they
do not appear to be used when using the downloaded clang.