Switch from tarball download to building inside the git repo The Extended Stable releases are not available in a tarball, so this is the only way to build this particular release branch.
Apart from the large fixed cost of cloning the giant multi-gigabyte
chromium repo, it just requires a few things such as setting up
depot tools and manually copying the debian folder.

As part of this change it's not helpful to have depot_tools pruned as
part of the build process, as it's called by some of the repo update
scripts. For the same reason, take tools/clang off the pruning list.