Remove gitsubreset subsection from debian/rules This is unneeded since we now copy the debian directory into the chromium git tree.
A command to perform a Submodule reset is already listed in the build
instructions, and a build error arises when not issuing a gitsubreset
command via debian/rules because the pruning script doesn't get patched.

The pruning script patch step and version override lines can safely be
moved to the setup section.