Build with PGO by default and add a setup_cast option Running debian/rules setup_cast enables building with chromecast enabled. The following patches and build options are respectively disabled and enabled :-
extra/ungoogled-chromium/fix-building-without-mdns-and-service-discovery.patch
extra/ungoogled-chromium/disable-dial-repeating-discovery.patch

enable_mdns=true
enable_service_discovery=true

It remains to be seen whether this will be sufficient for chromecast to work.

i cannot see any reason to not build with PGO since it has been officially
supported and seems to have been stable for a while now. In case of
PGO-related build errors, it can easily be disabled via a simple edit to the
rules file.