Build with use_system_freetype=true Although this is currently false by default, the chrome binary does depend on libfreetype (checked via ldd), presumably because of the action of the unbundling script that gets run during configuration.
However, pdf_bundle_freetype is false by default, and this appears
to tell the build system to look at the value of use_system_freetype
to determine whether to use system freetype (or not) when building
PDFium.

All in all, it seems to be a safer choicet to specify
use_system_freetype=true in order to avoid PDFium utilising the
bundled version of freetype.