Add a patch to raise the function import limits Since we are building with PGO and sample_profile_is_accurate=true, we can use the savings from the heavy size optimization of cold functions to more heavily optimize hot ones.
The basic import limit gets bumped from 5 to 10 and the hot function multiplier
is raised from 10 to 30. These numbers match those used to build the LTO-PGO
version of linux firefox, so I would assume they are a reasonable tradeoff
between size and performance.

Although not strictly a fair comparison due to different versions and building
with a reduced symbol_level of 0, -fdebug-types-section and AVX instructions
enabled, after building v101 the size of the chrome binary has increased by
5.64 Mib (3.1%) over that of v100.