Use import-instr-limit of 5 and remove sample_profile_is_accurate
Reading through build/config/compiler/BUILD.gn it seems that
sample_profile_is_accurate is not being used under linux.

The default function import values were 5 for import-instr-limit and 10 for
import-hot-multiplier. In a previous commit raised these to 10 and 30
respectively (as used by firefox).

However, an import-instr-limit of 5 amd import-hot-multiplier of 60 are
equivalent to 10 and 30 in terms of the limit for hot functions (10*30 = 5*60)
and reduces the size of the (previously released) binary by one megabyte.

The estimated increase over the size of a binary produced with default values
is now 2.5% versus 3.3% before, and should ensure that only frequently used
hot functions receive the speed benefit.

Finally, use '-mllvm' instead of '-plugin-opt' to set the hot multiplier in
accordance with how import-instr-limit is set elsewhere in the same file.