Add sample .gclient file This is for reference and potential future use, and is configured to prevent the download of several unneeded components including sysroot and nacl.
Note that it's possible via gclient to use a git shallow clone/fetch to
obtain/update the chromium source without pulling in gigabytes of history.
As of version 102.0.5005.61, a fully synced and prepared git clone is
22.7 GiB in size, versus 10.2 GiB for the tarball.

This size reduction lowers the relative disadvantage of git versus
downloading a tarball in terms of the extra storage requirement, but
overall git seems better supported, less problematic and more flexable.