Revert "Remove the clang download script command from debian/rules"
This is in preparation for adding back the ability to download and build
from the chromium tarball.

The clang/llvm script is now run as part of pgo_configure since bundled
(ie downloaded) clang/llvm is virtually a pre-requisite for building with
PGO (the clang version has to match that of PGO profile).

For those building with a chromium git checkout, both the clang and pgo
download scripts will have no effect, since both components are pulled
in during manual preparation.

This reverts commit 3bfc6cd490e27891f73d3d7ae617552dac07da28.