Add some granulated cleaning commands These are tuned to suit the situation, with a general clean before the main configuration and build processes, and a post-build/pre-packaging clean of the generated '__pycache__' directories.
These two commands are effectively combined in a single one under the
hardclean section to allow for a more thorough (manual) clean.

Besides not deleting the out/Release directory, this all pretty much follows
how things were before the recent changes to how cleaning is done, except
what gets deleted (and when this occurs) is a bit more customised.