Enable building without mDNS and Service Discovery At the time of writing, the enabling of mDNS and service discovery are a pull request.
This pull request changes the build flags to re-enable building these
two components, and a patch is added to allow them to be disabled via
a flag. An existing patch to fix building without the two services is
to be deleted.

This commit allows this to be reversed via the setup stage, by
re-purposing the (unfinished and unused) setup_cast debian/rules command.

One can run 'debian/rules setup_no_mdns' to set up the build, and the
deleted fix-building-without-mdns-and-service-discovery.patch will get
added back, and enable_mdns gets set to false in flags.gn.