Replace the command which deletes unbundled files The new method feeds build/linux/unbundle/remove_bundled_libraries.py with a list of directories to leave untouched.
This involves processing build/linux/unbundle/replace_gn_files.py to
extract a full list of unbundled components, and then remove each
member of $SYS_LIBS from it. This gives a list of components that
should be left untouched in the third_party directory.

These components are then used to look up the corresponding *.gn/gni
build file in replace_gn_files.py, and the get the list of directories
to give as arguments to replace_gn_files.py.

Tested on the command line, but not yet in a live build.