Add ten patches addenda/0001-hid-Handle-empty-input-reports.patch looks like it might be a potential security issue.
The first two in various/upstream are a continuation of the
move to selectively disable pointer compression [1], the
second of which mentions a benchmark gain of around 1.3%.

Patches three to eight fix issues discovered via running the
Speedometer 2.0 benchmark [2], and patch nine is a small
optimisation related to v8 pointer compression. Together
these ought to give a minor speed boost.

1/ https://bugs.chromium.org/p/chromium/issues/detail?id=1410145
2/ https://bugs.chromium.org/p/chromium/issues/detail?id=808503
3/ https://bugs.chromium.org/p/chromium/issues/detail?id=1325007#c65