Make the revertion of unbundling into a function command This is used in a couple of places to make the unbundling process behave in an orderly fashion.
I'm using a find command instead of the in-tree replace_gn_files.py
script with the '--undo' function, as restoration would fail in the
case of a library that had been removed from SYS_LIBS (via an edit
to d/rules) between an interruption to the build and the subsequent
rebuild.

The find command looks for anything ending with .gn.original or
.gni.original and does the restoration with no regard to whatever
libraries are incuded within the SYS_LIBS variable.

This commit also improves the find command by making it avoid dead
end paths like the out/Release directory. This should make things
a bit faster.