Update the hw-decoding-encoding runtime flag file After reading a chrome tracking bug [1] and an Arch Linux forum thread [2], I have come to the conclusion that enabling VAAPI on linux is something of a dark art.
It seems to work for some with various combinations of runtime
flags, environment variables, drivers and hardware, but not for
others.

The chromium bug is marked as fixed, which in my opinion is
something of an exaggeration. You have been warned.

Anyway, I have added a few flags, but these are not guaranteed
to work.

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=1236697
[2] https://bbs.archlinux.org/viewtopic.php?id=244031