Enable system jpeg by default in the configure script Most people will build for themselves, and probably build on the same distro that they intend to run under.
The only reason system jpeg was not the default is that this
would make my builds not run on say ubuntu (due to jpeg
package differences between the two distros).

If for some reason you are building under one distro and
running on another, then setting JPEG=0 is a safe choice.