Remove pgo_configure section from debian/rules PGO has long been a default option and two of its three lines are best located elsewhere.
Firstly, the LLVM_DIR variable and python script relate to
bundled clang. Support for this was removed a while back as
bundled clang was not polly-capable.

Secondly, the line invoking tools/update_pgo_profiles.py to
download the pgo profile only makes sense for tarball users
(git users have this file already). So move that to the
tarball section.

Thirdly, the pgo_data_path line is best placed next to the
other sed command which operates on the same file.