Fix/Improve the clang/llvm build instructions
Firsly, one needs to specify 'stage2-clang-bolt' as the
build target, and because bolt relies on intact debug
symbols, it's probably safer to not strip on install.

Add an import limit and hot multiplier to LDFLAGS like
we do in chromium.

Remove the 'git clean -dfx' step, as it doesn't appear
to actually clean anything.

I have also given details about pointing the system ld
towards lld via a symlink.