Simplify unbundling reversion search command
To undo unbundling we were searching the whole tree for files
which were named either '*.gn.original' or '*.gni.original'.

We were not searching for just '*.original' to avoid matching
any unrelated files (there is in fact a file with the
.original extension deep inside third_party/dawn).

An alternative is to patch the unbundling script to use a
more unique extension name. Using '.unbundled' instead of
'.original' allows us to just search for '*.bundled'.

Also, instead of searching the whole tree, restrict the
search to base directories of the files listed inside the
unbundling script (build/linux/unbundle/replace_gn_files.py).

Currently, there are only two such paths, third_party and
base/third_party.