Stop building chrome_crashpad_handler
This is related to crash reporting, which I understand is
disabled by upstream UC.

As such, I don't think the chrome_crashpad_handler binary
serves any function, and chromium seems to function fine
without it.

The binary itself is 1.2 MiB, and not building it possibly
makes the main chrome binary a bit smaller too.