Add a patch to disable vaapi av1 encoding
The debian libaom-dev package seems to lack a header
required by vaapi for av1 encoding, so we have to disable
this functionality to avoid build errors.

Since vaapi is an optional component, we might as well use
the configure script to unhide its patches.

Also, move various/fixes/vaapi-add-av1-support.patch to
the optional/system patch directory, so that the vaapi
patches are in the same place.