Use system xdg-mime and xdg-settings
The code - chrome/browser/shell_integration_linux.cc - as well
as the wrapper script deleted by the patch, seem to allow for
the possibility of using system xdg.

This adds a runtime dependency on xdg-utils, which most systems
will probably already have installed.

The native scripts are almost certainly newer, and perhaps
more complete, than the chromium ones (this is true on debian
stable).