toolchain: improve script and stop warning suppression
In the patching script We can avoid traversing the build
directory to make the find commands run quicker.

For brevity and clarity, we can call the find commands
via a function which has the filename as an argument.

For some reason the warning suppression clang flags do
not seem to be effective. The warnings appear during the
later build stages and basically aren't supressed via
patching of the source.