More runtime flag re-organisation
Since there are so many ui-related flags in miscellaneous, it
makes sense to put them into their own flag file (called ui).

We can get rid of debian/shims/chromium-flags.conf by moving
the '--disable-pings' flag into anti-tracking and the rest into
miscellaneous. (The '--no-pings' flag was recently put into
miscellaneous but was in fact already in chromium-flags.conf).

Also move the '--disable-webgl' flag from gpu-options to
anti-tracking, given the rationale for this flag includes
preventing info leakage.

I'm not sure if it's enabled by default, but as most people
will probably want VAAPI hardware decoding, we ought to have
the option in hw-decoding-encoding uncommented. Fix a minor
typo in that file too.

Finally add an option to isolation to allow site isolation
to be disabled, which could be useful for any websites broken
by this feature.