Add notes about enabling/troubleshooting vaapi
Also add a new flag '--enable-features=VaapiVideoDecodeLinuxGL' to
/etc/chromium.d/hw-decoding-encoding, which seems to a newer one
that someone on the ArchLinux forum has reported to work.

I have also re-organised hw-decoding-encoding, as a couple of the
flags probably no longer solve things like they might have done in
the past.

I found from running with an intel skylake under X11, the secret
sauce to enable vaapi was to install intel-media-va-driver-non-free
in conjuction with --disable-features=UseChromeOSDirectVideoDecoder.

Yes, that's the non-free version of intel-media-va-driver, which
seems to contain extra shaders amongst other things, and allows the
deprecated non-direct method to function.