d//configure.sh: Fix default value of LTO_JOBS
The problem here is that the script always enables the
patch, which in turn limits us to one thread.

If we change the default value of LTO_JOBS to 0 (and exclude
this value from the case statement), then the patch won't
get applied and we'll end up using '-Wl,--thinlto-jobs=all'
in our ldflags.

From a comment in the code, a value of '0' is broadly
equivalent to the upstream default of 'all', so no more
nastly surprises.