Update d/configure.sh and d/rules for bookworm release
So many libraries are now new enough to unbundle under
the newly-released debian bookworm, that I've had to
alter the script to enable them for both stable and
unstable. Only dav1d is exclusive to unstable.

In debian/rules, rather than moving the newly unbundled
libraries to the vertically-listed section, I have kept
them in a separate block to make things clearer (ie they
will likely be the ones which will become unstable-only
over time).

The png libraries goes into the vertical list as I
presume that it is unlikely to age between debian stable
releases, such that it needs to become an unstable-only
option.

You can now build without openh264 via running the
configure script with 'OPENH264=0'. The default is to
enable the it and unbundle its system library.
I'll presume that this library, which wasn't in stable
before bookworm was released, is not going to age like
some of the other ones surely will.

I have decided not to move patches or rename folders
in the case that a folder called unstable contains a
patch which can potentially apply to stable. As stable
versions become too old you would need to revert such
changes, raising the chance of errors for little gain.
It's far easier to flip things around in the script
and leasve the patches where they are.

Building debs with multi-threaded xz compression is now
possible under stable (XZ_THREADED=1).