Update d/configure.sh and d/rules for new debian stable
I've compiled a fresh copy of clang/llvm after doing
a full dist-upgrade of my stable chroot, and have now
got stuck in to the new debian stable release.

So many libraries are now new enough to unbundle that
I've had to move certain things outside of the
'if [ UNSTABLE -eq 1 ]' section of the configure script
and create a little section to accomodate this new
reality.

As stable distributions age and their libraries become
too old, parts of this new section (which applies to
both stable and unstable) can be moved back to the
unstable section.

Various other bits and pieces have been changed and the
rules file updated in tandem with configure.sh.