Allow enabling WebGPU
This is not being enabled by default for the reasons
below.

The build flag for WebGPU (aka dawn) seems to have been
enabled by default in v114, but doesn't seem to be
functional on a test build.

I suspect this is merely for the convenience of developers
and possibly to try and catch build issues. One source I
read claims that only Windows and MAC are supported, and
this seems to accord with the skia_use_dawn build flag only
being enabled by default on these platforms.

Whether enabling it or not, I have disabled the swiftshader
component of WebGPU in debian rules, in common with the
pre-existing disablement of swiftshader elsewhere.

A runtime flag in /etc/chromium.d/gpu-options needs to be
uncommented to enable WebGPU in an enabled build.