configure.sh: Only guard the install file from modification
The other files under the 'Modify debian directory files'
heading near the end of the script always (or practically
always) get modified.

I don't think we need to be ultra-conservative about only
touching files when necessary to do so, as the aggregation
of the sed commands into one single command already
provides a huge win in this respect.