Stop the translate-script-url unsupported flag warning
This is the little pop-up warning message that appears
under the address bar.

Since we are using the same url (via a flag file) that is used
internally by chromium, this shouldn't be an issue.

The concern mentioned in the code is about PII (Personally
Identifying Information), and this could be seen as an
admission by google that they have the potential to personally
identify its users via their use of translate.

Note this only affects those building with translate enabled.