Enable google translate by default
The only caveat here is that the translate-script-url runtime
flag needs to be uncommented in /etc/chromium.d/google-translate.

The translation script url is the same one used internally by
chromium [1,2] (the internal url is scrambled by the domain
substitution regardless of whether the browser is compiled with
translate enabled or not).

Any translation script can be a vector of personally identifying
information (PII) according to a comment in the source [3].

Therefore if this url is not supplied then users should at most
only be exposed to the following urls :-

https://redirector.gvt1.com/edgedl/chrome/dict/
https://translate.googleapis.com/

These are taken from the patch which re-enables translate (it
reverts the urls from being blocked and logged). The first is
assigned to the kDownloadServerUrl variable, and the second is
assigned to kSecurityOrigin.

The first I'm not 100% sure is strictly necessary to enable
translate (but most likely is, as I recall testing it quite
extensively a few years ago), and looks like a dictionary
download url, used either locally or remotely.

The second, as noted in the code itself, seems to be where the
privacy issue lies. It appears to me that this is the part that
can read (and potentially spy on) whatever webpage text you are
translating.

The bottom line is that if this translation script url isn't
enabled (via the flag file), then there shouldn't be ant real
privacy issue.

I cannot guarantee this 100%, and if anyone is paranoid about
the issue then they can compile with translate disabled.

Lastly, move both translate patches to optional/translate
and to better describe what translate-reverse-enable.patch
actually does, rename it to translate-revert-disable.patch.

Also, move the google-translate flag file alongside all the
others in debian/etc/chromium.d and alter the install file
to install it by default.

[1] https://translate.googleapis.com/translate_a/element.js
[2] components/translate/core/browser/translate_script.cc
[3] chrome/browser/ui/startup/bad_flags_prompt.cc