Stop installing icudtl.dat by default
This should mean that we don't have to touch the install file
in most cases, as none of the defaults employ the INS variable.

Only paranoid builds disabling translate, those disabling Qt
and those installing the policies file (which can be installed
permanently by hand), would alter the install file.