Re-work MARCH/MTUNE handling
This commit should ensure basic sanitisation of the MARCH/MTUNE
variables in the case that one is set and one isn't.

The rules are :-

MARCH matches MTUNE if MTUNE is set and MARCH is not specified.
MARCH will not match MTUNE if MTUNE is set to generic.

MTUNE matches MARCH if MARCH is set and MTUNE is not specified.
MTUNE will be set to generic if MARCH is one of the x86-64 levels.

MARCH and MTUNE will reset to their default values [1] if MARCH
is (mis-)set to generic. Overall the most common mistakes and
misconfigurations should be avoided.

MARCH and MTUNE are printed if the sanitisation process changes
any of their values, and their respective patches will only be
touched when non-default values are used [1].

[1] Defaults are MARCH=x86-64-v2 and MTUNE=generic