Re-work polly optimisation patches
First of all, we now only apply the vectorization and
extra optimisation to the -O3 levels in accordance to
what is generally recommended by many sources.

I remember reading somewhere (one of the mailing lists
iirc) that the documentation needed to be overhauled
and is possibly out of date. It does mention that -O1
and -O2 are permitted, but let's play safe and avoid
over-optimising, which could be counter-productive.

The polly parallel feature should really be regarded as
experimental, as it's problematic with javascript. It
needs to be implemented in the right places.

For anyone who wants to experiment, the polly-parallel
patch has been altered to add a config to the main
build file (build/config/compiler/BUILD.gn), which can
be called from other BUILD.gn files.

A new configure script variable POLLY_THREADS is
introduced. When set to 2 or greater, it enables a patch
(optional/polly-parallel-threads.patch) which sets the
'-polly-num-threads=<N>' clang flag. By default N=2.

If polly parallel is enabled then a libgomp runtime
dependency now gets enabled in d/control. The handling
of deps_enable is altered to handle runtime dependencies
which lack the '-dev' suffix.

Finally, the polly patch names have been shortened by
removing the llvm prefix.