Stop post-compilation removal of __pycache__ folders
It can take a bit of time for the find command to scan the
whole source tree for '__pycache__' folders, and since we
are not making a source tarball, we can skip this step.

These folders (very slightly) speed up rebuilds, and only
really need to be removed when updating the source tree to
a new release (via running git clean -dfx).

I have kept the associated rules command, such that
removal can be invoked manually by running :-

./debian/rules find_remove_pycache