Stop routine cleaning before compilation
As mentioned in a recent commit, we don't need to delete
'__pycache__' folders, as we don't make a tarball from our
source trees.

After further thought, I've decided that we should extend
this to not doing any routine cleaning at all.

We are forced into running a clean in only two situations.
First, we need to 'git clean -dfx' to clean when updating
the repo to a new release. Second, when rebuilding after a
completed build we are forced to omit the '-nc' switch
when running dpkg-buildpackage (which invokes dh_clean).

Outside of the above two cleaning scenarios, running these
'find and delete' commands serves no real purpose and just
wastes time, cpu cycles and disk activity. Anything that
would get deleted is harmless cruft which doesn't affect
the build process.

After this change, the d/rules file loses eight lines and
is now a bit simpler and easier to follow. The hardclean
rules command remains for anyone who feels the need to
use it.