Disable system brotli
There seems to be a new version of brotli with a new
header (brotli/shared_dictionary.h) that the debian
development package doesn't seem to have.

We can just use the bundled version until a solution
is found. I have made a small change to debian/rules
that can be wired into the configure script at a
future time.