d/configure.sh: Make testing the script easier
Testing the script outside of a build tree would cause
errors and an unnecessary warning about a directory not
existing.

To solve this we test for the existence of a top-level
'third_party' directory to see whether we are inside a
build tree or not. If we are outside of a build tree,
then treat things like a test run and suppress the
errors and warning.

A test run prevents the proper PGO profile path being
written into d/submodules/ungoogled-chromium/flags.gn
and stops debian/devutils/update_launcher.sh for being
executed (which in turn stops d/shims/chromium being
generated).

Not executing debian/devutils/update_launcher.sh is
useful if testing the script on a partition without
executable perimissions. (The whole command can howver
be run manually via prepending 'sh' to it).