configure.sh: Add RTC_AVX2 and re-jig V8_AVX2
These both build and run on non-AVX2 machines so it's
probably easier to enable both by default whilst giving
the ability to disable (which probably saves a tiny amount
of binary size).

Since V8_AVX2 is now set to 1 by default, we may as well
make it independent of the AVX2 variable, and give the
user the ability to override the default values of all
three avx2-related variables.