Fix unbundled libusb and enable by default
I wasn't sure whether third_party/libusb was actually
used by the build or not, but on further inspection it
appears that it's probably used at build-time, but not
at runtime.

I'm therefore going to unconditionally enable unbundling
it, which entails removing all traces of it from the
configure script (which was broken anyway).

We also add optional/system/libusb-api.patch, which
might well be needed to fix a potential api issue.