Add widevine-cdm/widevine-cdm.sh
This is a script which will download the latest chrome
deb file from google and build a deb out of the extracted
Widevine CDM files.

The widevine-locations patch needs to be altered to
accommodate the location used by the widevine-cdm package
(/usr/lib/WidevineCdm). The $HOME/.local/lib path is
retained as a fallback, so that existing setups continue
to function.

Full instructions are contained in widevine-cdm/README.md.

Note: for copyright/licencing reasons the widevine-cdm
cannot be distributed.