Add optional/fixes/extensions-in-tools-menu.patch
This takes the extensions menu out of the main program
menu and puts it back into the 'More tools' menu. This
is where it resided prior to version 116.

A new configure script variable EXT_TOOLS_MENU controls
whether to include the above patch. The default is 1,
to include the patch.