Make enablement of skia-gamma.patch configurable
I've read at least one comment that fonts seem to
be a bit too light, especially thin/light ones.

This is something that I have noticed myself, but
wasn't too bothered by.

I suspect that various/fixes/skia-gamma.patch might
be to blame. This raises the gamma exponent used
for text blitting from 1.2 to 2.2, which would
indeed lighten fonts.

I'm guessing that this patch was perhaps motivated
because the author was using an older less bright
display, or possibly for reasons of personal taste.

Nobody else I'm aware of uses this patch, so the
best solution seems to be to make this patch
optional.

Pass SKIA_GAMMA=1 to configure.sh to enable it.