Remove chromium-launcher submodule
The chromium-launcher upstream hasn't been updated in two
years [1] and the shell launcher works more than well
enough, being transparent, editable by users and perhaps
less prone to security issues.

The only reason the submode wasn't removed was simply that
there wasn't a great need to. Now it just gets in the way,
so blitz it.

[1] https://github.com/foutrelis/chromium-launcher/commits/master