Drop bromite gateway attack related patches
This has caused lots of issues in the past, and seems to me
to be a bit too complex and intrusive, for dubious benefit.

It requires another large patch which reverts the removal
of an upstream commit, and just seems to me to be more
likely to introduces issues rather than solve them.

Dropped :-

bromite/Block-gateway-attacks-via-websockets.patch
bromite/Revert-removal-of-execution-context-address-space.patch