configure.sh: Autodetect locally installed clang
The problem was that BUNDLED_CLANG was set to 0 by default
which would cause things to fail if there was no locally
installed clang.

With this commit, those opting to use bundled clang should
no longer have to set BUNDLED_CLANG=1.

System clang (ie via a deb package) is not supported [*],
so the choices are either locally installed (self-built)
or the copy bundled with chromium.

We now set BUNDLED_CLANG to one by default, but set it to
zero if a locally installed clang is detected inside
/usr/local. However, we allow the value to be forced to
respect users who customise their setup.

[*] System clang (specifically debian's) is in constant
need of patches to fix build errors, and most often does
not support PGO due to it being an older version than the
bundled version. (Supporting it would be a headache, and
is best left to users who can eg patch their rules file
and possibly jump through various other hoops.)