Build with bundled icu
Upstream chromium now requires icu version 73, which
is only available in debian experimental. This also
prevents unbundling libxml and libxslt due to a
dependency chain between the three libraries.

Iit probably won't be long until icu moves to unstable,
so we ought to set things up from a perspective of
system icu being enabled by default, but gated to
disabled until this move happens.

In the meantime flip the SYS_ICU default in configure.sh
to zero. Those on unstable with icu installed from
experimental can configure with SYS_ICU=1.

It's unlikely that system icu will be an option for
stable for a long time, so force SYS_ICU=0 if STABLE=1
to try and stop stable builders shooting themselves in
the foot.

Finally, fix the dependency situation in configure.sh
and debian/control.in. Remove libxml from the former
since it wasn't in the latter. Add libicu-dev to
both, with a version check in control.in to help
prevent version mismatch build errors.