configure.sh: Make processing of dependencies more robust
The sed command was having no effect when attempting to
disable a library in control.in, when the library name
lacked a preceding space.

These sorts of formatting 'errors' are easily made when
editing control.in. They are only errors in the sense
that the configuration script might not function in a
correct fashion (the actual syntaxt is correct purely
from the point of view of a debian control file).

Fix these sorts of issues by employing liberal instances
of '[ ]*' in the dependency-related sed commands to
allow for there being zero or any number of spaces
around the library string.

This change should make enabling/disabling dependencies
much more 'fault' tolerant.