Re-work clang/llvm self-compilation advice
Building a PGO/BOLT-enabled toolchain is a multi-stage
process, where the compiler is built two or three times
in different instances.

It seems that the optimisations we put in with the
optimise.sh script were lost after the first stage and
were therefore ineffective. This is no huge loss.

We can turn this to our advantage by simplifying the
whole process. We remove the optimisation script and
simplify the build steps to make the whole thing less
complex and scary.

The rest of the commit is just a re-wording of the
instructions to try and make things a bit clearer for
those interested in compiling the browser themselves.