configure.sh: Bail out if LTO_DIR does not exist
This is the safer choice, rather than letting other
commands run afterwards.

This makes it possible to fix up the cause of the
error and re-run the script without any worries or
extra hassles.

Move the relevant LTO stuff up the script to a
position before any sed commands get executed.