configure.sh: Enable a couple of patches by default
Here we avoid more needless enabling or disabling of
patches. Try and touch the series file only when a
non-default choice is made.

System libaom and non-bundled clang are defaults.

Also get rid of the other usage of the system_patches
variable when adding system/unstable/ to the dav1d
patch. This method was originally intended to be
used with multiple patches.