configure.sh: Simplify things by removing polly parallel
Since learning that polly-parallelisation is an unsafe
optimisation (and not really effective either) it's time
to relegate this.

The two polly-parallel patches, plus the requisite
structured binding patch have been moved out of the
debian directory to a new attic folder in the repo
root, just in case someone finds them useful.

This commit removes a net amount of thirty lines and
three variables from the configuration script, plus
three patches from the series file, which greatly
simplifes things.