configure.sh: Enable libaom, libavif and dav1d by default
Of libaom, libavif and dav1d, the first two are enabled
without conditions for both stable and unstable.

The default position of the script is that unstable is
the default, so twiddle things only when STABLE=1.

We now have more items in the control, rules and series
files which are set to the default, hence the script in
most cases doesn't need to do as much flipping.

It's probably easier to view these files and form a
rough idea of what the defaults are.