Reorganise system libraries and their dependency chains
Trying to switch to bundled libwebp led me down the rabbit hole
of unbundling again, which led to many more changes than I
originally envisaged.

Here are some new things I have discovered (or had forgotten) :-

libavif depends on libaom and dav1d
libaom depends on libyuv
fontconfig depends on freetype
freetype depends on brotli and libpng
libtiff depends on libwebp

libyuv cannot be unbundled because it depends on libstdc++
which in turn is incompatible with control flow integrity.

I prefer to to avoid a situation where the internal bundled
library is loaded together with an external version of it.
Besides the duplication, there's a potential mixing of versions
which doesn't seem like a wise idea.

To avoid this situation we switch to the bundled versions of
libaom and libavif to prevent a 'hidden' dependency on libyuv.

Likewise we switch to bundled versions of fontconfig and
freetype to avoid a 'hidden' dependency on brotli.

I have refreshed, renamed and re-enabled the dav1d patch
below, which is required when building with bundled libavif :-

optional/system/unstable/dav1d-no-unbundled-libavif.patch
renamed to...
optional/system/unstable/dav1d-bundled-libavif.patch

The now redundant optional/system/unstable/libaom-headers.patch
has been moved to the attic directory.

To account for the situation with bundled brotli and prepare
for a future need to disable freetype, I have added SYS_FREETYPE
as a configure script variable and fixed up deps_disable to
allow disabling dependencies like brotli which are not of the
'*-dev' form.

In d/rules use_system_freetype is actually true by default, and
must be set to false as a pre-requisite to disabling system
freetype. To completely disable system freetype one additionally
needs to compile with pdf_bundle_freetype=false.