configure.sh: Avoid modifying d/patches/series.debian
This commit avoids an unneeded write to series.debian.

Instead of doing an in-place edit of series.debian using
the sed '-i' option, we use sed (without this option) to
store the modified content of this file in a variable.

We now effectively have series.debian as a string, and we
similarly turn the content of the UC submodule series file
into a string (via using a cat command). All we need to do
to produce the final series file is to echo the two strings.