Abort all unbundling steps if domsubcache.tar.gz exists
The general advice for altering things mid-build, eg changes
to patches, is to revert domain substitution before making
changes. Not doing so risks domain substitution noticing
that target files have been changed when attempting to
restore the original files from domsubcache.tar.gz and
erroring out. Patches cannot then be unapplied cleanly.

To avoid the possibility that unbundling changes files
within domsubcache.tar.gz, we need to test whether this
cache file exists before allowing any unbundling commands
to run. I'm not aware of this having happened yet, but
this seems a wise move to make sure that it doesn't.

Finally, in the override_dh_prep section of d/rules. we
need to reverse the order of undo_unbundling and the
reversion of domain substitution so that the former
is performed after the latter.

A nice side effect of this is that resuming a build should
be a bit faster due to not having to run a find command to
search through the contents of the base/third_party/ and
third_party/ directories.