Move en-US.pak instead of copying and deleting it
The override_dh_auto_build section of debian/rule is not
really an appropriate place to perform this copy command
which is related to installation.

It's also illogical to copy and delete a source file
rather than just moving it.

So let's replace the copy and delete with one move
command under override_dh_auto_install-indep.